AD-HOC feat (Apache): Explicitly deny dotfiles
In the previous build the application was not able to request dotfiles
as they did not exist, and the handler was not set up to handle them
while the application was not installed.

However, this commit introduces configuration to explicitly block
dotfiles, thereby clearing up any ambiguity as to whether they're
appropriately hidden.