AD-HOC feat (docker): Allow persisting data outside application
Snipe appears to store various kinds of data outside the application
itself, and in some sort of persistence directories.

This commit builds on top of the Dockerfile bundled in the application
proper to include these various stateful directories in the docker
definition.

== Design Notes ==

=== Separate Volumes ===

Initially this was implemented as a single volume, with symlinks to the
appropriate places in the application directory. However, docker appears
to drop all content below the VOLUME declared directory. Additionally,
where users are initialising volumes themselves these required
directories will not be present.

Many orchestration tools allow representing multiple different volumes
on a single disk, with multiple bind mounts. The author feels this is a
better abstraction to represent volumes.